# openapi.yaml
openapi: 3.0.3
info:
  title: Post-it Tasks API
  version: 1.0.0
  description: API for managing post-it style reminder tasks
servers:
  - url: http://localhost:8080
    description: Local development
paths:
  /api/tasks:
    get:
      tags: [Tasks]
      summary: List all tasks
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      tags: [Tasks]
      summary: Create a task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /api/tasks/{id}:
    get:
      tags: [Tasks]
      summary: Get a task by id
      parameters:
        - $ref: "#/components/parameters/TaskId"
      responses:
        "200":
          description: Task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "404":
          description: Not found
    put:
      tags: [Tasks]
      summary: Update a task
      parameters:
        - $ref: "#/components/parameters/TaskId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "404":
          description: Not found
    delete:
      tags: [Tasks]
      summary: Delete a task
      parameters:
        - $ref: "#/components/parameters/TaskId"
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
components:
  parameters:
    TaskId:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
  schemas:
    Task:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        content:
          type: string
        color:
          type: string
          enum: [yellow, pink, mint, blue, peach]
        posX:
          type: integer
        posY:
          type: integer
        done:
          type: boolean
        createdAt:
          type: string
          format: date-time
    TaskRequest:
      type: object
      properties:
        title:
          type: string
          maxLength: 120
        content:
          type: string
          maxLength: 2000
        color:
          type: string
        posX:
          type: integer
        posY:
          type: integer
        done:
          type: boolean